Replace SpeciesWithAttributes
Replace all usages of SpeciesWithAttributes with UiSpeciesAttributes.
Add method to the ObservationRepository to get by list of ids.
Move lettercode mapping into TemplateService
Remove SpeciesWithAttributes entity and repository.

+++ api/src/main/java/au/org/aodn/nrmn/restapi/repository/projections/LetterCodeMapping.java
@@ -2,7 +2,7 @@ package au.org.aodn.nrmn.restapi.repository.projections;
 
 public interface LetterCodeMapping {
 
-    Integer getObservableItemId();
+    Long getObservableItemId();
 
     String getLetterCode();
