+++ api/src/main/java/au/org/aodn/nrmn/restapi/repository/projections/SpeciesWithAttributesCsvRow.java
@@ -19,5 +19,5 @@ public class SpeciesWithAttributesCsvRow {
 
     private Double l95;
 
-    private Integer lMax;
+    private Long lMax;
 }
